+++ client/cody-jetbrains/src/main/java/com/sourcegraph/cody/vscode/TextDocument.java
@@ -1,0 +1,17 @@
+package com.sourcegraph.cody.vscode;
+
+import java.net.URI;
+
+public interface TextDocument {
+  URI uri();
+
+  String fileName();
+
+  int offsetAt(Position position);
+
+  String getText();
+
+  String getText(Range range);
+
+  Position positionAt(int offset);
+}