+++ client/cody/src/completions/completion.test.ts
@@ -5,72 +5,34 @@ import {
     CompletionResponse,
 } from '@sourcegraph/cody-shared/src/sourcegraph-api/completions/types'
 
-import { CodyCompletionItemProvider, __test_only_resetCache } from '.'
+import { mockVSCodeExports } from '../testSetup/vscode'
 
-jest.mock('vscode', () => {
-    class Position {
-        public line: number
-        public character: number
+import { CodyCompletionItemProvider, inlineCompletionsCache } from '.'
 
-        constructor(line: number, character: number) {
-            this.line = line
-            this.character = character
-        }
-    }
-    class Range {
-        public start: Position
-        public end: Position
-
-        constructor(start: Position, end: Position) {
-            if (typeof start === 'number') {
-                throw new TypeError('this version of the constructor is not implemented')
+jest.mock('vscode', () => ({
+    ...mockVSCodeExports(),
+    InlineCompletionTriggerKind: {
+        Invoke: 0,
+        Automatic: 1,
+    },
+    workspace: {
+        getConfiguration() {
+            return {
+                get(key: string) {
+                    switch (key) {
+                        case 'cody.debug.filter':
+                            return '.*'
+                        default:
+                            return ''
+                    }
+                },
             }
-            this.start = start
-            this.end = end
-        }
-    }
-    class InlineCompletionItem {
-        public content: string
-        constructor(content: string) {
-            this.content = content
-        }
-    }
-    return {
-        Position,
-        Range,
-        InlineCompletionItem,
-        InlineCompletionTriggerKind: {
-            Invoke: 0,
-            Automatic: 1,
         },
-        workspace: {
-            getConfiguration() {
-                return {
-                    get(key: string) {
-                        switch (key) {
-                            case 'cody.debug.filter':
-                                return '.*'
-                            default:
-                                return ''
-                        }
-                    },
-                }
-            },
-            onDidChangeTextDocument() {
-                return null
-            },
+        onDidChangeTextDocument() {
+            return null
         },
-        window: {
-            createOutputChannel() {
-                return null
-            },
-            showErrorMessage(message: string) {
-                throw new Error(message)
-            },
-            activeTextEditor: { document: { uri: { scheme: 'not-cody' } } },
-        },
-    }
-})
+    },
+}))
 
 function createCompletionResponse(completion: string): CompletionResponse {
     return {
@@ -79,6 +41,8 @@ function createCompletionResponse(completion: string): CompletionResponse {
     }
 }
 
+const CURSOR_MARKER = '<cursor>'
+
 /**
  * A test helper to trigger a completion request. The code example must include
  * a pipe character to denote the current cursor position.
@@ -86,7 +50,7 @@ function createCompletionResponse(completion: string): CompletionResponse {
  * @example
  *   complete(`
  * async function foo() {
- *   |
+ *   ${CURSOR_MARKER}
  * }`)
  */
 async function complete(
@@ -129,12 +93,12 @@ async function complete(
         true // disable timeouts
     )
 
-    if (!code.includes('|')) {
+    if (!code.includes(CURSOR_MARKER)) {
         throw new Error('The test code must include a | to denote the cursor position')
     }
 
-    const prefix = code.slice(0, code.indexOf('|'))
-    const suffix = code.slice(code.indexOf('|') + 1)
+    const prefix = code.slice(0, code.indexOf(CURSOR_MARKER))
+    const suffix = code.slice(code.indexOf(CURSOR_MARKER) + CURSOR_MARKER.length)
 
     const codeWithoutCursor = prefix + suffix
 
@@ -200,10 +164,10 @@ function truncateMultilineString(string: string): string {
 }
 
 describe('Cody completions', () => {
-    beforeEach(() => __test_only_resetCache())
+    beforeEach(() => inlineCompletionsCache.clear())
 
     it('uses a simple prompt for small files', async () => {
-        const { requests } = await complete('foo |')
+        const { requests } = await complete(`foo ${CURSOR_MARKER}`)
 
         expect(requests).toHaveLength(3)
         expect(requests[0]!.messages).toMatchInlineSnapshot(`
@@ -234,7 +198,7 @@ describe('Cody completions', () => {
                 public end: Position
 
                 constructor(startLine: number, startCharacter: number, endLine: number, endCharacter: number) {
-                    this.startLine = |
+                    this.startLine = ${CURSOR_MARKER}
                     this.startCharacter = startCharacter
                     this.endLine = endLine
                     this.endCharacter = endCharacter
@@ -262,12 +226,12 @@ describe('Cody completions', () => {
     })
 
     it('does not make a request when in the middle of a word', async () => {
-        const { requests } = await complete('foo|')
+        const { requests } = await complete(`foo${CURSOR_MARKER}`)
         expect(requests).toHaveLength(0)
     })
 
     it('completes a single-line at the end of a sentence', async () => {
-        const { completions } = await complete('foo = |', [
+        const { completions } = await complete(`foo = ${CURSOR_MARKER}`, [
             createCompletionResponse("'bar'"),
             createCompletionResponse("'baz'"),
         ])
@@ -275,17 +239,17 @@ describe('Cody completions', () => {
         expect(completions).toMatchInlineSnapshot(`
             Array [
               InlineCompletionItem {
-                "content": "'bar'",
+                "insertText": "'bar'",
               },
               InlineCompletionItem {
-                "content": "'baz'",
+                "insertText": "'baz'",
               },
             ]
         `)
     })
 
     it('completes a single-line at the middle of a sentence', async () => {
-        const { completions } = await complete('function bubbleSort(|)', [
+        const { completions } = await complete(`function bubbleSort(${CURSOR_MARKER})`, [
             createCompletionResponse('array) {'),
             createCompletionResponse('items) {'),
         ])
@@ -293,17 +257,17 @@ describe('Cody completions', () => {
         expect(completions).toMatchInlineSnapshot(`
             Array [
               InlineCompletionItem {
-                "content": "array) {",
+                "insertText": "array) {",
               },
               InlineCompletionItem {
-                "content": "items) {",
+                "insertText": "items) {",
               },
             ]
         `)
     })
 
     it('does not make a request when context has a selectedCompletionInfo', async () => {
-        const { requests } = await complete('foo = |', undefined, undefined, {
+        const { requests } = await complete(`foo = ${CURSOR_MARKER}`, undefined, undefined, {
             selectedCompletionInfo: {
                 range: {} as any,
                 text: 'something',
@@ -315,7 +279,7 @@ describe('Cody completions', () => {
     })
 
     it('filters out completions that start with whitespace', async () => {
-        const { completions } = await complete('function bubbleSort(|)', [
+        const { completions } = await complete(`function bubbleSort(${CURSOR_MARKER})`, [
             createCompletionResponse('\t\t\tarray) {'),
             createCompletionResponse('items) {'),
         ])
@@ -323,7 +287,7 @@ describe('Cody completions', () => {
         expect(completions).toMatchInlineSnapshot(`
             Array [
               InlineCompletionItem {
-                "content": "items) {",
+                "insertText": "items) {",
               },
             ]
         `)
@@ -331,27 +295,48 @@ describe('Cody completions', () => {
 
     describe('odd indentation', () => {
         it('filters our odd indentation in single-line completions', async () => {
-            const { completions } = await complete('const foo = |', [createCompletionResponse(' 1')])
+            const { completions } = await complete(`const foo = ${CURSOR_MARKER}`, [createCompletionResponse(' 1')])
 
             expect(completions).toMatchInlineSnapshot(`
                 Array [
                   InlineCompletionItem {
-                    "content": "1",
+                    "insertText": "1",
                   },
                   InlineCompletionItem {
-                    "content": "",
+                    "insertText": "",
                   },
                 ]
             `)
         })
 
-        it.todo('removes odd indentation in multi-line completions')
-        it.todo('handles \t in multi-line completions')
+        it('removes odd indentation in multi-line completions', async () => {
+            const { completions } = await complete(
+                `
+                function test() {
+                    ${CURSOR_MARKER}
+                }`,
+                [createCompletionResponse(' foo()\n     bar()')]
+            )
+
+            expect(completions[0].insertText).toBe('foo()\n    bar()')
+        })
+
+        it('handles \t in multi-line completions', async () => {
+            const { completions } = await complete(
+                `
+                function test() {
+                \t${CURSOR_MARKER}
+                }`,
+                [createCompletionResponse(' foo()\n\t bar()')]
+            )
+
+            expect(completions[0].insertText).toBe('foo()\n\tbar()')
+        })
     })
 
     describe('multi-line completions', () => {
         it('triggers a multi-line completion at the start of a block', async () => {
-            const { requests } = await complete('function bubbleSort() {\n  |')
+            const { requests } = await complete(`function bubbleSort() {\n  ${CURSOR_MARKER}`)
 
             expect(requests).toHaveLength(3)
             expect(requests[0]!.stopSequences).not.toContain('\n')
@@ -362,7 +347,7 @@ describe('Cody completions', () => {
                 `
                 class Foo {
                     constructor() {
-                        |
+                        ${CURSOR_MARKER}
                     }
                 }`,
                 [
@@ -388,15 +373,15 @@ describe('Cody completions', () => {
             expect(completions).toMatchInlineSnapshot(`
                 Array [
                   InlineCompletionItem {
-                    "content": "if (foo) {
+                    "insertText": "if (foo) {
                             console.log('foo1');
                         }",
                   },
                   InlineCompletionItem {
-                    "content": "console.log('foo')",
+                    "insertText": "console.log('foo')",
                   },
                   InlineCompletionItem {
-                    "content": "",
+                    "insertText": "",
                   },
                 ]
             `)
@@ -407,7 +392,7 @@ describe('Cody completions', () => {
                 `
                 class Foo {
                     constructor() {
-                        |
+                        ${CURSOR_MARKER}
                     }
                 }`,
                 [
@@ -422,7 +407,7 @@ describe('Cody completions', () => {
 
             expect(completions[0]).toMatchInlineSnapshot(`
                 InlineCompletionItem {
-                  "content": "console.log('foo')
+                  "insertText": "console.log('foo')
 
                         console.log('bar')
 
@@ -435,7 +420,7 @@ describe('Cody completions', () => {
             const { completions } = await complete(
                 `
                 if (check) {
-                    |
+                    ${CURSOR_MARKER}
                 }`,
                 [
                     createCompletionResponse(`
@@ -448,7 +433,7 @@ describe('Cody completions', () => {
 
             expect(completions[0]).toMatchInlineSnapshot(`
                 InlineCompletionItem {
-                  "content": "console.log('one')
+                  "insertText": "console.log('one')
                 } else {
                     console.log('two')",
                 }
@@ -459,7 +444,7 @@ describe('Cody completions', () => {
             const { completions } = await complete(
                 `
                 if (check) {
-                    |
+                    ${CURSOR_MARKER}
                 `,
                 [
                     createCompletionResponse(`
@@ -470,7 +455,7 @@ describe('Cody completions', () => {
 
             expect(completions[0]).toMatchInlineSnapshot(`
                 InlineCompletionItem {
-                  "content": "console.log('one')
+                  "insertText": "console.log('one')
                 }",
                 }
             `)
@@ -480,7 +465,7 @@ describe('Cody completions', () => {
             const { completions } = await complete(
                 `
                 function myFunction() {
-                    |
+                    ${CURSOR_MARKER}
                     console.log('three')
                 }
                 `,
@@ -496,7 +481,7 @@ describe('Cody completions', () => {
 
             expect(completions[0]).toMatchInlineSnapshot(`
                 InlineCompletionItem {
-                  "content": "console.log('one')
+                  "insertText": "console.log('one')
                     console.log('two')",
                 }
             `)
@@ -506,7 +491,7 @@ describe('Cody completions', () => {
             const { completions } = await complete(
                 `
                 function test() {
-                    |`,
+                    ${CURSOR_MARKER}`,
                 [
                     createCompletionResponse(`
                     console.log('foo')
@@ -527,22 +512,49 @@ describe('Cody completions', () => {
             expect(completions).toMatchInlineSnapshot(`
                 Array [
                   InlineCompletionItem {
-                    "content": "console.log('foo')
+                    "insertText": "console.log('foo')
                     console.log('foo')
                     console.log('foo')
                     console.log('foo')
                     console.log('foo')",
                   },
                   InlineCompletionItem {
-                    "content": "console.log('foo')",
+                    "insertText": "console.log('foo')",
                   },
                   InlineCompletionItem {
-                    "content": "console.log('foo')",
+                    "insertText": "console.log('foo')",
                   },
                 ]
             `)
         })
 
-        it.todo('handles tab/newline interop in completion truncation')
+        it('handles tab/newline interop in completion truncation', async () => {
+            const { completions } = await complete(
+                `
+                class Foo {
+                    constructor() {
+                        ${CURSOR_MARKER}`,
+                [
+                    createCompletionResponse(`
+                    console.log('foo')
+                    \t\tif (yes) {
+                    \t\t    sure()
+                    \t\t}
+                    \t}
+
+                    \tadd() {
+                        \tconsole.log('bar')
+                        }`),
+                ]
+            )
+
+            expect(completions[0].insertText).toMatchInlineSnapshot(`
+                "console.log('foo')
+                \t\tif (yes) {
+                \t\t    sure()
+                \t\t}
+                \t}"
+            `)
+        })
     })
 })