+++ client/cody/CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to Sourcegraph Cody will be documented in this file.
 
 ### Fixed
 
+- Cody completions: Fixed interop between spaces and tabs. [pull/52497](https://github.com/sourcegraph/sourcegraph/pull/52497)
 - Fixes an issue where new conversations did not bring the chat into the foreground. [pull/52363](https://github.com/sourcegraph/sourcegraph/pull/52363)
 
 ### Changed